requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.10.15
Django==2.1.3
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
pytz==2018.7
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
wincertstore==0.2